+++ CHANGELOG.md
@@ -17,8 +17,9 @@
 - Migrate `ContactPatientBottomSheet` to a `BaseBottomSheet`
 - Add Material motion in `ContactPatientBottomSheet` for handling view visibility changes
 - Migrate `ScheduleAppointmentSheet` to a `BaseBottomSheet`
+- Migrate `PatientsTabScreen` to a fragment
 - [In Progress: 20 Jan 2021] Material Theme-ing Migration
-- [In Progress: 08 Feb 2021] Migrate app to use ViewBinding
+- [In Progress: 05 Apr 2021] Migrate app to use ViewBinding
 
 ### Changes
 - Add Somali language option to language selection